Move consensus to own crate
Also move some of the shared definitions (Point, Peer) to the
ouroboros crate in a new 'protocol' module.
