+++ crates/amaru/src/lib.rs
@@ -17,8 +17,3 @@
 /// The sync pipeline is responsible for fetching blocks from the upstream node and
 /// applying them to the local chain.
 pub mod sync;
-
-/// Consensus interface
-///
-/// The consensus interface is responsible for validating block headers.
-pub mod consensus;
