+++ crates/consensus/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "amaru-consensus"
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+async-trait.workspace = true
+gasket.workspace = true
+hex.workspace = true
+indicatif.workspace = true
+indoc.workspace = true
+miette.workspace = true
+pallas-codec.workspace = true
+pallas-crypto.workspace = true
+pallas-math.workspace = true
+pallas-network.workspace = true
+pallas-primitives.workspace = true
+pallas-traverse.workspace = true
+rocksdb.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+ouroboros = { path = "../ouroboros" }
+amaru-ledger = { path = "../ledger" }
+
+[dev-dependencies]
+envpath.workspace = true
+insta.workspace = true
+proptest.workspace = true
+rand.workspace = true
+tempfile.workspace = true
