+++ crates/consensus/src/lib.rs
@@ -10,11 +10,9 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License.
+// limitations under the License
 
+/// Consensus interface
+///
+/// The consensus interface is responsible for validating block headers.
 pub mod consensus;
-pub mod kes;
-pub mod ledger;
-pub mod protocol;
-pub mod validator;
-pub mod vrf;
