+++ crates/ouroboros/Cargo.toml
@@ -10,11 +10,13 @@ mockall.workspace = true
 pallas-codec.workspace = true
 pallas-crypto.workspace = true
 pallas-math.workspace = true
+pallas-network.workspace = true
 pallas-primitives.workspace = true
 pallas-traverse.workspace = true
 rayon.workspace = true
 serde.workspace = true
 thiserror.workspace = true
+tokio.workspace = true
 tracing.workspace = true
 vrf_dalek.workspace = true
